chore: separate core & CLI, unify DB path, auto-migrate on start
