+++ drizzle.config.ts
@@ -1,11 +1,12 @@
 import { defineConfig } from "drizzle-kit";
-import { DB_PATH } from "./src/db/path";
 
 export default defineConfig({
-  schema: "./src/db/schema/*",
+  schema: "./src/core/db/schema/*",
   out: "./drizzle",
   dialect: "sqlite",
   dbCredentials: {
-    url: DB_PATH,
+    url: "./intents.db",
   },
+  strict: true,
+  verbose: true,
 });
