+++ src/core/db/index.ts
@@ -1,7 +1,11 @@
+import path from "node:path";
 import Database from "better-sqlite3";
 import { drizzle } from "drizzle-orm/better-sqlite3";
-import { DB_PATH } from "./path";
+import { migrate } from "drizzle-orm/better-sqlite3/migrator";
+
 import * as schema from "./schema";
+
+export const DB_PATH = path.resolve(process.cwd(), "intents.db");
 
 const sqlite = new Database(DB_PATH);
 
@@ -9,3 +13,5 @@ export const db = drizzle(sqlite, {
   schema,
   logger: process.env.NODE_ENV === "development",
 });
+
+migrate(db, { migrationsFolder: "./drizzle" });
